requirements.txt
flask
requests
openai
sqlalchemy
schedule
python-dotenv
flask_sqlalchemy
